refactor: extract dropdown menu into OptimizationPopover component
